requirements.txt
fastapi
uvicorn
python-dotenv
bs4
selenium